refactor: extract inline HTML parser
